Enhance code structure for calling OMDB API
